댓글 구현중..

+++ TwoLineProject/src/test/java/com/hyunwoo/mapper/ReplyMapperTest.java
@@ -50,7 +50,7 @@ public class ReplyMapperTest {
 	
 	@Test
 	public void testDelete() {
-		int targetRno = 1;
+		Long targetRno = 1L;
 		
 		mapper.delete(targetRno);
 	}
